+++ database/migrations/2022_10_22_153911_create_warranties_table.php
@@ -1,0 +1,38 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create('warranties', function (Blueprint $table) {
+            $table->id();
+            $table->string('model', 100);
+            $table->string('serial', 100)->unique();
+            $table->string('warranty_no', 100)->unique();
+            $table->date('purchase_date');
+            $table->boolean('type')->default(0); //0:out, 1:in
+            $table->foreignId('customer')->nullable()->constrained('customers')->onDelete('restrict');
+            $table->timestamps();
+            $table->softDeletes();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::dropIfExists('warranties');
+    }
+};