+++ database/migrations/2022_10_22_162839_alter_warranty_no_and_purchase_date_in_warranties_table.php
@@ -1,0 +1,34 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('warranties', function (Blueprint $table) {
+            $table->dropUnique(['warranty_no']);
+            $table->string('warranty_no', 100)->nullable()->change();
+            $table->date('purchase_date')->nullable()->change();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('warranties', function (Blueprint $table) {
+            //
+        });
+    }
+};