+++ resources/views/layout/customer/show.blade.php
@@ -1,0 +1,5 @@
+<span class="text-default">{{ $data->name}}</span>
+<br><span>{{ $data->phone}}</span>
+@if($data->address)
+    <br><small class="text-muted">{{ $data->address}}</small>
+@endif