+++ resources/views/layout/footer.blade.php
@@ -1,4 +1,4 @@
-<footer class="text-center p-3 border-top">
-    <h6 class="font-logo text-default">PANA<span class="text-primary">TECH</span><span class="text-secondary">.</span></h6>
+<footer class="text-center p-3 pt-4">
+    <h3 class="font-logo text-default">PANA<span class="text-primary">TECH</span><span class="text-secondary">.</span></h3>
     <small class="text-muted">&#169; 2022 Erwinzilla Labs. All rights reserved.</small>
 </footer>